[FEAT] Finished the first version
The Chain of Responsibility and Strategy patterns are implemented.

Everything seems to be working fine.

What's left is to test all possibilites. If no bug is detected, I may leave this code as how it is now and call it a day.

The next commit may be pushed after I theoretically catch an error that doesn't follow the rules of the game. Besides, if a unexpected error which is not related to the rules shows up, I might upload a FIX commit as well.